Time updates
Created a ConvertTime class to hold the conversion methods. I'm at the point now that
the time is checked for hours of options and that the start time is before the end time. The time is converted to GMT before being saved into the database. I need to set up the time so that when I pull it from the database it is converted from GMT time to the local time of the pc for display.