Final updates
Replaced the version in fxml docs with 11 instead of the 15.0.1 to get ride of the warning message when changing between scene. Removed the customer combo box and using table view instead. User can now just click on the customer from the table view. Also changed filter option when deleting an appointment it no longer puts all appointments in the table but only the ones for that customer.